# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_chainlaunch_chainlaunch_pkg_notifications_http.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  http.CreateProviderRequest:
    properties:
      config: {}
      isDefault:
        type: boolean
      name:
        maxLength: 255
        minLength: 1
        type: string
      notifyBackupFailure:
        type: boolean
      notifyBackupSuccess:
        type: boolean
      notifyNodeDowntime:
        type: boolean
      notifyS3ConnIssue:
        type: boolean
      type:
        allOf:
        - $ref: '#/definitions/notifications.ProviderType'
        enum:
        - SMTP
    required:
    - config
    - name
    - type
    type: object
...
  http.ProviderResponse:
    properties:
      config: {}
      createdAt:
        type: string
      id:
        type: integer
      isDefault:
        type: boolean
      lastTestAt:
        type: string
      lastTestMessage:
        type: string
      lastTestStatus:
        type: string
      name:
        type: string
      notifyBackupFailure:
        type: boolean
      notifyBackupSuccess:
        type: boolean
      notifyNodeDowntime:
        type: boolean
      notifyS3ConnIssue:
        type: boolean
      type:
        $ref: '#/definitions/notifications.ProviderType'
      updatedAt:
        type: string
    type: object
...
  http.TestProviderRequest:
    properties:
      testEmail:
        type: string
    required:
    - testEmail
    type: object
  http.TestProviderResponse:
    properties:
      message:
        type: string
      status:
        type: string
      testedAt:
        type: string
    type: object
...
  http.UpdateProviderRequest:
    properties:
      config: {}
      isDefault:
        type: boolean
      name:
        maxLength: 255
        minLength: 1
        type: string
      notifyBackupFailure:
        type: boolean
      notifyBackupSuccess:
        type: boolean
      notifyNodeDowntime:
        type: boolean
      notifyS3ConnIssue:
        type: boolean
      type:
        allOf:
        - $ref: '#/definitions/notifications.ProviderType'
        enum:
        - SMTP
    required:
    - config
    - name
    - type
    type: object
...
  notifications.ProviderType:
    enum:
    - SMTP
    type: string
    x-enum-varnames:
    - ProviderTypeSMTP
...
host: localhost:8100
info:
  contact:
    email: support@chainlaunch.com
    name: API Support
    url: http://www.chainlaunch.com/support
  description: ChainLaunch API provides services for managing blockchain networks
    and cryptographic keys
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ChainLaunch API
  version: "1.0"
paths:
...
  /notifications/providers:
    get:
      consumes:
      - application/json
      description: Get a list of all notification providers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.ProviderResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_chainlaunch_chainlaunch_pkg_notifications_http.ErrorResponse'
      summary: List notification providers
      tags:
      - notifications
    post:
      consumes:
      - application/json
      description: Create a new notification provider with the specified configuration
      parameters:
      - description: Provider creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.CreateProviderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.ProviderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_chainlaunch_chainlaunch_pkg_notifications_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_chainlaunch_chainlaunch_pkg_notifications_http.ErrorResponse'
      summary: Create a notification provider
      tags:
      - notifications
  /notifications/providers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a notification provider
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_chainlaunch_chainlaunch_pkg_notifications_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_chainlaunch_chainlaunch_pkg_notifications_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_chainlaunch_chainlaunch_pkg_notifications_http.ErrorResponse'
      summary: Delete a notification provider
      tags:
      - notifications
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific notification provider
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.ProviderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_chainlaunch_chainlaunch_pkg_notifications_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_chainlaunch_chainlaunch_pkg_notifications_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_chainlaunch_chainlaunch_pkg_notifications_http.ErrorResponse'
      summary: Get a notification provider
      tags:
      - notifications
    put:
      consumes:
      - application/json
      description: Update an existing notification provider with new configuration
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: integer
      - description: Provider update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.UpdateProviderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.ProviderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_chainlaunch_chainlaunch_pkg_notifications_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_chainlaunch_chainlaunch_pkg_notifications_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_chainlaunch_chainlaunch_pkg_notifications_http.ErrorResponse'
      summary: Update a notification provider
      tags:
      - notifications
  /notifications/providers/{id}/test:
    post:
      consumes:
      - application/json
      description: Test a notification provider
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: integer
      - description: Test provider request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.TestProviderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.TestProviderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_chainlaunch_chainlaunch_pkg_notifications_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_chainlaunch_chainlaunch_pkg_notifications_http.ErrorResponse'
      summary: Test a notification provider
      tags:
      - notifications
...
schemes:
- http
- https
securityDefinitions:
  BasicAuth:
    type: basic
  CookieAuth:
    in: cookie
    name: session_id
    type: apiKey
swagger: "2.0"
...